refactor(holdings): move PriceChartData + PortfolioBreakdownData to types/beancounter
Domain payload types now live in `types/beancounter.d.ts` alongside the
existing `QuickSellData`, `WeightClickData`, `SetCashBalanceData`, etc.
set, so `CardView`, `Rows`, `aggregated.tsx`, and `[code].tsx` no longer
import callback shapes across component boundaries.

PortfolioBreakdownPopup test now uses `makeAsset` + a new
`makePortfolioBreakdown` builder from `src/test-fixtures/beancounter.ts`
instead of hand-rolling Asset / PortfolioBreakdown literals.

Skipped CodeRabbit's `types/beancounter` → `@types/beancounter` nitpicks:
TS6137 reserves the `@types/*` prefix for DefinitelyTyped, so the alias
is dead for `.d.ts` files; the bare path via `baseUrl` is the repo
convention.
